Refactoring snapshot naming.
