--- src/Snapshooter/ISnapshotFileInfoResolver.cs
@@ -1,11 +1,0 @@
-﻿using Snapshooter.Core;
-
-namespace Snapshooter
-{
-    public interface ISnapshotFullNameResolver
-    {
-        SnapshotFullName ResolveSnapshotFullName();
-        SnapshotFullName ResolveSnapshotFullName(string snapshotName);
-        SnapshotFullName ResolveSnapshotFullName(string snapshotName, string nameExtension);
-    }
-}
